Don't ignore inline comments
Partially fixes corrupting DEF files when patching them (#10), but
also introduces line breaks for every comment. Unfortunate, but
it's better than making the file unreadable by Ikemen and there
doesn't seem to be a way to do both with this INI library.

Also adds a quick quality of life improvement to the console output.